Fix if then

+++ _deploy/cibuild.sh
@@ -1,9 +1,8 @@
 #!/usr/bin/env bash
 set -e # halt script on error
 
-if [ -f .cache/.jekyll-metadata ] then
-  cp .cache/.jekyll-metadata .
-fi
+# Copy metadata if cached
+[ -f .cache/.jekyll-metadata ] && cp .cache/.jekyll-metadata .
 
 bundle exec jekyll build --incremental
 bundle exec htmlproofer ./_site --disable-external --allow-hash-href --empty-alt-ignore --internal-domains adrianba.net
